@@ -4,7 +4,7 @@
  * DESC
  *
  * @license Use of this software requires acceptance of the License Agreement. See LICENSE file.
- * @copyright Copyright � 2016-present Heidelberger Payment GmbH. All rights reserved.
+ * @copyright Copyright � 2016-present heidelpay GmbH. All rights reserved.
  * @link https://dev.heidelpay.de/JTL
  * @author Ronja Wann
  * @category JTL
